Refactor BridgeSupportTest class, move flyover tests into another test class. Add tests for check minimum for pegin in flyover tx. Add tests for fix parsing of segwit outputs for pegins in flyover tx. Add tests for add check minimum for pegin per UTXOs.
